% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocplot.R
\name{calibplot}
\alias{calibplot}
\title{Make a smoothed calibration plot}
\usage{
calibplot(
  pred,
  negrug,
  posrug,
  ideal,
  ylim = c(0, 1),
  xlim = c(0, 1),
  capuci = TRUE,
  xlabel = "Predicted probability of presence",
  ylabel = "Probability of presence",
  filename = NULL,
  title = "Calibration plot"
)
}
\arguments{
\item{pred}{Data frame. Predicted probability of presence at occurrence and background locations}

\item{negrug}{Numeric. Model prediction of presence at background points.}

\item{posrug}{Numeric. Model prediction of presence at occurrence points.}

\item{ideal}{Function.}

\item{ylim}{Numeric. Two-element vector giving the x-axis for the plot; default is c(0, 1).}

\item{xlim}{Numeric. Two-element vector giving the y-axis for the plot; default is c(0, 1).}

\item{capuci}{Logical. Should the upper confidence limit of the smoothed probability of presence by clamped to 1?}

\item{xlabel}{Character. Label to be applied to the x-axis.}

\item{ylabel}{Character. Label to be applied to the y-axis.}

\item{filename}{Character. Full path to the file into which the graph will be plotted. Default of NULL plots to the standard graphics device.}

\item{title}{Character. Title for the plot. Default is a generic "Calibration plot".}
}
\value{
A ggplot2 object
}
\description{
Make a smoothed calibration plot for a presence-only model using the method by Phillips and Elith
}
\details{
{
This is an adaptation of code published by Phillips and Elith (2010. POC plots: calibrating species distribution models with presence-only data. Ecology 91:2476–2484).

Although it is possible to call this function with user-generated parameter values, it is ideally designed to be called by the function \link{POCplot}.

This implementation uses ggplot2 to render the plot.

}
}
\examples{
\dontrun{}
}
