% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePR.R
\name{computePRcurve}
\alias{computePRcurve}
\title{Compute Precision-Recall curve}
\usage{
computePRcurve(thisModel, occSWD, bkgSWD, responseType = "cloglog")
}
\arguments{
\item{thisModel}{maxnet object. A maxnet object representing the fitted model}

\item{occSWD}{data.frame. Environmental data for presence or occurrence records in MaxEnt SWD format}

\item{bkgSWD}{data.frame. Environmental data for background locations in MaxEnt SWD format}

\item{responseType}{Character. A MaxEnt response scale; one of "link", "exponential", "logistic" or "cloglog". Default is "cloglog"}
}
\value{
A PR curve object produced by the function \emph{pr.curve} in the package \pkg{PRROC}
}
\description{
Compute a Precision-Recall (PR) curve and AUC value for a maxnet model
}
