% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_summary.R
\name{feature_summary}
\alias{feature_summary}
\title{Generate a feature summary table}
\usage{
feature_summary(model_files = "")
}
\arguments{
\item{model_files}{Character array. Array of file names (including full path) to a set of maxnet models}
}
\value{
A data.frame with variables 'min", 'max' and 'votes' and row names representing variable and feature names, 'votes' is the number of replicate model fits which selected each feature.
}
\description{
Generate a feature summary table
}
