% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMaxnet.R
\docType{package}
\name{fitMaxnet}
\alias{fitMaxnet}
\title{\pkg{fitMaxnet}}
\description{
A package of functions to provide methods for preparing data files and
fitting MaxEnt models using the function maxnet in the R-package maxnet.
}
