% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModels.R
\name{fitModels}
\alias{fitModels}
\title{Fit a set of MaxEnt models using the maxnet package}
\usage{
fitModels(
  taxonName = NULL,
  occPath = NULL,
  backgroundPath = NULL,
  nBkgPoints = 5000,
  baseOutputPath = NULL,
  replTags = "",
  multSet = 1,
  outputType = "link"
)
}
\arguments{
\item{taxonName}{The taxonomic name associated with a fitted model.}

\item{occPath}{File system path to the file holding occurrence records and associated environmental data.}

\item{backgroundPath}{File system path to a file holding a matrix of environmental data representing the background environment for the supplied occurrence data.}

\item{nBkgPoints}{Number (integer) giving the number of background points to sub-sample from data in the file pointed to by \emph{backgroundPath}.}

\item{baseOutputPath}{File system path to the base folder to be used for model output. A sub-folder with the taxonomic name supplied in \emph{taxonName}. It will be created if it doesn't already exist to receive output.}

\item{replTags}{A character array of tags representing replicates to be fitted.}

\item{multSet}{A numeric array of values to be used for overall regularisation multipliers. At least one value must be present which may be the default value of 1.}

\item{outputType}{Character. Output scaling of fitted model: "link" is raw linear predictor scores, while "exponential", "logistic" and "cloglog" generate a non-linear re-scaling of the raw linear predictor scores. See \link[maxnet]{maxnet} for further information.}
}
\value{
Nothing
}
\description{
A set of MaxEnt models will be fitted using the maxnet function in the package maxnet. A model will be fitted for each combination of replTags and multSet values.
}
\examples{
\dontrun{}
}
