% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kFoldSet.R
\name{makeFolds}
\alias{makeFolds}
\title{Make k sets of indices to samples for k-fold cross-validation}
\usage{
makeFolds(numSamples, k)
}
\arguments{
\item{numSamples}{Integer. The number of samples to be partitioned into k folds.}

\item{k}{Integer. The number of folds.}
}
\value{
A list of numeric vectors of length k, where the elements of each vector are the indices of the samples to be included on that fold.
}
\description{
The function randomly allocates sample indices to k numeric vectors. The length of each vector is set initially to be the nearest whole number which, when multiplied by the number of folds, comes as close as possible to the number of samples. Any residual of un-allocated sample indices remaining after the initial allocation are distributed randomly to the k groups. Each selected group receives only 1 additional sample index.
}
