% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskExtrapolation.R
\name{maskExtrapolation}
\alias{maskExtrapolation}
\title{Mask a projection raster removing extrapolation areas}
\usage{
maskExtrapolation(
  maxnetModel,
  envPath,
  projRas,
  maskOutpath = dirname(projRas),
  fileLabel,
  makePlots = TRUE
)
}
\arguments{
\item{maxnetModel}{Full path to a maxnet model fitted by \link{fit_maxnet} and saved as an .Rd file.}

\item{envPath}{Path to a set of environmental rasters showing the environment onto which the model has been projected.}

\item{projRas}{Raster of the projected maxnet model.}

\item{maskOutpath}{Full path to the folder into which the mask raster and projection raster with mask applied will be written}

\item{fileLabel}{Label to be used to distinguish the filename of the saved mask raster.}

\item{makePlots.}{Logical. Make basic plots of raster objects for review and quality control. Default is TRUE.}
}
\value{
Nothing but has side-effect of writing two raster files: mask raster and projection raster with mask applied.
}
\description{
Compute and apply a mask of areas within which strict extrapolation has been detected
}
\details{
{
Some info on method + details of coding used in the output raster...

Note method for forming an output name for the mask raster
}
}
