% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePR.R
\name{plotPRcurve}
\alias{plotPRcurve}
\title{Plot a Precision-Recall (PR) Curve}
\usage{
plotPRcurve(PRobj, plotFilename = NULL)
}
\arguments{
\item{PRobj}{pr.curve object. A PR ROC object with type = "PR" produced by the function \link[PRROC]{pr.curve} in the package \pkg{PRROC} run with parameter curve = TRUE}

\item{plotFilename}{Character (string). Optional file name to save the plot output (as PNG formatted graphics file)}
}
\value{
A ggplot2 object
}
\description{
Plot a Precision-Recall (PR) Curve
}
