% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocplot.R
\name{POCplot}
\alias{POCplot}
\title{Presence-only smoothed calibration plot}
\usage{
POCplot(
  pred,
  back,
  linearize = FALSE,
  capUpperValues = TRUE,
  title = "Calibration plot",
  filename = NULL
)
}
\arguments{
\item{pred}{Numeric. Vector of model predictions at occurrence points.}

\item{back}{Numeric. Vector of model predictions at background points.}

\item{linearize}{Logical. Should the logistic transform be applied to model predictions? The default is FALSE; see Details.}

\item{capUpperValues}{Logical. Should computed values in smoother fit be clamped to 1?}

\item{filename}{Character. Full path to a file into which the calibration plot will be written. Default (NULL) will cause the plot to appear on the standard graphics device.}

\item{title.}{Character. Title to appear in the plot. Default is a generic "Calibration plot".}
}
\value{
Data frame of results from the smoothing process
}
\description{
Produce a smoothed calibration plot for a presence-only model
}
\details{
{
This is an adaptation of code published by Phillips and Elith (2010. POC plots: calibrating species distribution models with presence-only data. Ecology 91:2476–2484).

The original script set the default value of the parameter \emph{linearize} to TRUE which applies a logistic transformation to the y-values in the POCplot. In development trials, this frequently produced numerical failures when values approaching 1 where transformed. It is perhaps intended to transform 'raw' values supplied by fitted ENMs to the logistic scale. However, the most frequent output from MaxEnt and MaxEnt-like models is already transformed onto a logistic or complementary log-log ('cloglog') scale. Inadvertent double application of a linearising transformation will cause numeric failures and severely distorted calibration plots. So, the default for \emph{linearize} is set to FALSE.

}
}
