% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectMaxnet.R
\name{projectMaxnet}
\alias{projectMaxnet}
\title{Project maxnet model}
\usage{
projectMaxnet(
  taxonName = NULL,
  maxnetModel,
  type = "exponential",
  doClamp = FALSE,
  dataType = "FLT4S",
  baseOutputPath,
  fileLabel = NULL,
  makeTaxonFolder = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{taxonName}{String. Taxonomic name associated with this model; used to construct a file name for the output raster file.}

\item{maxnetModel}{String. \emph{Full} path name to the .Rd file storing a fitted maxnet model produced by the companion functions \link[fitMaxnet]{fit_maxnet} and \link[fitMaxnet]{fitModels}.}

\item{type}{String. The type of scaling applied to predicted model values.}

\item{doClamp}{Logical. Should values of predictors (covariates) be clamped to those seen during model fitting? Default is FASLE, no clamping.}

\item{dataType}{String. Data format to be used. Default is "FLT4S". Only alternative at moment is "INT4S".}

\item{baseOutputPath}{String. Full path to the output folder to receive the produced raster.}

\item{fileLabel}{String. An identifying tag to be included in the output filename.}

\item{makeTaxonFolder}{Logical. Should a sub-folder on \emph{baseOutputPath} be created using \emph{taxonName}?}

\item{quiet}{Logical. Proceed without emitting progress messages?}
}
\value{
Nothing
}
\description{
Project a maxnet model fitted using fitModel
}
\details{
{
Function \link[fitMaxnet]{prepProjData} \strong{MUST} be run before calling this function to prepare global data objects needed for the projection.

\strong{NOTE:} The resulting raster is written to the specified output folder with the file name composed by concatenating the taxon name spaces replaced by underscores), "projection" and, if supplied, the character object \emph{fileTag}.

\strong{If a file with this name already exists in the output folder, it will be overwritten.}

If dataType = "INT4S" is given, output is multiplied by 1,000 and truncated to an integer. Therefore, real values (to 4 dec. places) must be regenerated after loading the TIFF file by dividing the raster values by 1,000.
}
}
