% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleBackground.R
\name{sampleBackground}
\alias{sampleBackground}
\title{Select background cells}
\usage{
sampleBackground(
  occData,
  baseRaster,
  boundsPolygon,
  nBkgSamples = maxBkgSamples,
  maxBkgSamples = 10000,
  trace = FALSE
)
}
\arguments{
\item{occData}{a data frame or matrix with at least two numeric columns which can be associated with longitude (or X) and latitude (or Y)}

\item{baseRaster}{A raster layer object representing the geometry of the environmental data layers to be used for modelling}

\item{boundsPolygon}{An sf or SpatialPolygons* object defning the boundary within which background points must be selected}

\item{nBkgSamples}{Numeric. The number of samples to constitute the background set with the default = maxBkgSamples}

\item{maxBkgSamples}{Numeric. Maximum number of background samples to be generated (default = 10,000)
#' @param trace Logical. Emits additional diagnostic messages when TRUE; default is FALSE}
}
\value{
A integer array of indices to selected background cells on the base raster
}
\description{
Select a set of background cells from a base raster layer
}
\details{
{The number of background samples may need to be adjusted to suit particular situations.

The default value for maxBkgSamples matches the value used in the original MaxEnt Java implementation (see Phillips, S. J., and M. Dudík. 2008. Modeling of species distributions with Maxent: new extensions and a comprehensive evaluation. Ecography 31:161–175. for the reason it was selected).

For many taxa, sampling a constrained area means that less than 10,000 cells will be available. When this occurs, the user is warned and the number of background samples returned is the maximum physically available within the specified buffer.

}
}
\examples{
\dontrun{}
}
