% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleBackground.R
\name{sampleBackgroundOld}
\alias{sampleBackgroundOld}
\title{Perform background cell selection}
\usage{
sampleBackgroundOld(
  occupiedCells,
  nSamples = 1000,
  baseMap,
  method = "unconstrained",
  diskRadius = 200
)
}
\arguments{
\item{occupiedCells}{Gridcell indices of occupied cells on the raster defined in \emph{baseMap}.}

\item{nSamples}{Number fo samples to be drawn to represent the background.}

\item{baseMap}{A RasterLayer object defining the grid to be sampled.}

\item{method}{Character string selecting which of two methods to use to draw the background sample.}

\item{diskRadius}{Radius of disk in kilometers around each occurrence cell for "Union of Disks' method. Ignored otherwise.}
}
\value{
An integer array giving the indices of selected grid cells on the raster defined in \emph{baseMap}
}
\description{
Given a set of occupied grid cells on a specified raster, perform background cell selection using one of two methods: unconstrained, or 'union of disks' based on VanDerWal et al. (2009).
}
