% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleBackground.R
\name{unionOfDisks}
\alias{unionOfDisks}
\title{Selected background cells from disks of given radius}
\usage{
unionOfDisks(baseMap, seedCells, diskRadius = 200)
}
\arguments{
\item{baseMap}{A RasterLayer object representing the spatial grid to be sampled}

\item{seedCells}{Cell indices on the \emph{baseMap} grid representing occupied cells to act as seed cells}

\item{diskRadius}{Radius in kilometres of the disk to be used fro background cell sleection}
}
\value{
An integer array of selected grid cells on the \emph{baseMap} raster
}
\description{
Perform background cell selection from a union of disks of specified radius (in km) around each occurrence cell. Based on VanDerWal et al. (2009).
}
