% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImportance.R
\name{varImportance}
\alias{varImportance}
\title{Variable importance for a maxnet model}
\usage{
varImportance(
  theModel,
  occSWD,
  bkgSWD,
  responseType = c("link", "exponential", "cloglog", "logistic"),
  numReplicates = 5
)
}
\arguments{
\item{theModel}{Object of class maxnet}

\item{occSWD}{Data.frame. Environmental data at occurrence locations in SWD format.}

\item{bkgSWD}{Data.frame. Background environmental data in SWD format.}

\item{responseType}{Character. A MaxEnt response scale; one of "link", "exponential", "logistic" or "cloglog".}

\item{numReplicates}{Integer. Number of permutations performed to compute importance.}
}
\value{
A named vector of percent importance scores for all variables present in the maxnet model object.
}
\description{
Variable importance for a maxnet model
}
