% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boyce_index.R
\name{boyce}
\alias{boyce}
\title{Boyce's continuous index}
\usage{
boyce(fit, obs, nclass = 0, window.w = "default", res = 100, PEplot = FALSE)
}
\arguments{
\item{fit}{A vector or Raster-Layer containing the predicted suitability values}

\item{obs}{A vector containing the predicted suitability values or xy-coordinates (if fit is a Raster-Layer) of the validation points (presence records)}

\item{nclass}{Number of classes or vector with classes threshold. If nclass=0, Boyce index is calculated with a moving window (see next parameters)}

\item{window.w}{width of the moving window (by default 1/10 of the suitability range)}

\item{res}{resolution of the moving window (by default 100 focals)}

\item{PEplot}{if TRUE, plot the predicted to expected ratio along the suitability class}
}
\value{
A named list...
}
\description{
Boyce's continuous index
}
\examples{
\dontrun{}
}
