% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bufferPoints.R
\name{bufferPoints}
\alias{bufferPoints}
\title{Make buffer around occurrence points}
\usage{
bufferPoints(occ_pts, bufferDist_km = 200)
}
\arguments{
\item{occ_pts}{A data frame or matrix with at least two columns representing latitude and longitude of occurrence records}

\item{bufferDist_km}{Size of buffer to be built around the occurrence points in kilometres. Default value of 200 km follows the advice of VanDerWal et al. (2009). See details below.}
}
\value{
A spatial polygon object from the package \emph{sp}
}
\description{
Produce a spatial polygon object representing the boundary within which background points will be sampled for ENM fitting
}
\details{
{
\emph{bufferPoints()} implements the widely used simple circular buffering approach introduced to niche modelling by VanDerWal et al. 2009. 'Selecting pseudo-absence data for presence-only distribution modeling: How far should you stray from what you know?' \emph{Ecological Modelling} 220:589–594.
The default value of 200 km for the buffer distance is the value recommended by VanDerWal et al. (2009) for a range of taxa which they considered in their paper. Naturally, the value to be used is a matter for consideration by the user. If there is a clear indication from available information regarding a taxon's mode of dispersal, or there is guidance from population genetics data, then another value might be supplied for this parameter.
This implementation uses the spatial processing resources provided by the packages \emph{sp}, \emph{sf} and \emph{rgeos} to achieve good performance and accurate handling of spatial data.
}
}
\note{
{
Currently, this function can only process \bold{occurrence data from the Australian continent}. Providing global coverage is feasible and will be implemented shortly for the OzWeeds project.

The function only outputs a spatial polygon object projected on the WGS84 datum. If you need to use the buffer polygon on other projections, please reproject using, say, \emph{spTransform()} from the package \emph{sp}.
}
}
\examples{
\dontrun{}
}
