% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bufferPoints_old.R
\name{bufferPoints_old}
\alias{bufferPoints_old}
\title{Make buffer around occurrence points}
\usage{
bufferPoints_old(
  occ_pts,
  bufferDist_km = 200,
  X_col = NULL,
  Y_col = NULL,
  epsg = NULL,
  trace = FALSE
)
}
\arguments{
\item{occ_pts}{A data frame or matrix with at least two columns representing longitude and latitude (X and Y) of occurrence records}

\item{bufferDist_km}{Numeric. Size of buffer to be built around the occurrence points in kilometres. Default value of 200 km follows the advice of VanDerWal et al. (2009). See details below.}

\item{X_col}{Integer. Optionally given the column index of the X-coordinate.}

\item{Y_col}{Integer. Optionally given the column index of the Y-coordinate.}

\item{epsg}{Integer. Optionally specify the EPSG integer code for the projection. If not specified, an attempt is made to determine the CRS from the X and Y coordinate values.}

\item{trace}{Logical. Should messages be emitted to assist progress tracking or debugging}
}
\value{
A spatial polygon object from the package \emph{sf}
}
\description{
Produce a spatial polygon object representing the boundary within which background points will be sampled for ENM fitting
}
\details{
{
\emph{bufferPoints()} implements the widely used simple circular buffering approach introduced to niche modelling by VanDerWal et al. 2009. 'Selecting pseudo-absence data for presence-only distribution modeling: How far should you stray from what you know?' \emph{Ecological Modelling} 220:589–594.
The default value of 200 km for the buffer distance is the value recommended by VanDerWal et al. (2009) for a range of taxa which they considered in their paper. Naturally, the value to be used is a matter for consideration by the user. If there is a clear indication from available information regarding a taxon's mode of dispersal, or there is guidance from population genetics data, then another value might be supplied for this parameter.
This implementation uses the spatial processing resources provided by the packages \emph{sp}, \emph{sf} and \emph{rgeos} to achieve good performance and accurate handling of spatial data.
}
}
\note{
{
Currently, this function can only process \bold{occurrence data from the Australian continent}. Providing global coverage is feasible and will be implemented shortly for the OzWeeds project.

The function outputs a simple features ('sf') polygon object on the same projection which was inferred when \emph{occ_pts} was parsed. That is, either EPSG:4326 (WGS84) if 'longitude' and 'latitude' column names were found, or EPSG:3577 (Australian ALber's Equal Area) if 'X' and 'Y' column names were found. If you need to use the buffer polygon on other projections, please re-project the returned object using, say, \emph{st_transform()} from the package \emph{sf}.
}
}
\examples{
\dontrun{}
}
