% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeMESS.R
\name{computeMESS}
\alias{computeMESS}
\title{Compute MESS rasters}
\usage{
computeMESS(
  thisModel = NULL,
  envPath = NULL,
  occSWD,
  bkgSWD,
  varImpThreshold = 0,
  outPath = "",
  MESSonly = TRUE,
  ...
)
}
\arguments{
\item{thisModel}{A maxnet model object}

\item{envPath}{Full specified path to the environmental predictor raster layers used to fit the maxnet model}

\item{occSWD}{numeric matrix. Reference data; typically created by concatenating occurrence SWD and background SWD files used to fit the model}

\item{bkgSWD}{numeric matrix}

\item{varImpThreshold}{Numeric. Value used to filter variables by their importance scores; percentage importance therefore ranges from 0 to 100.}

\item{outPath}{Character. Path to folder into which output raster layers will be written}

\item{MESSonly}{Logical. Compute only the MESS raster layer? Default is TRUE. If FALSE, then MESS component rasters for each variable plus the MESS raster layer are computed}

\item{...}{Optional parameters passed to the raster function writeRaster}
}
\value{
Stuff like what is produced by dismo::mess()
}
\description{
Compute set of MESS rasters for a maxnet model
}
\details{
A MESS computation is performed blah blah.

To facilitate the processing of very large, high-resolution rasters, only those variables which are used in the final model are included in the computation of MESS output.
}
\examples{
\dontrun{}

}
