% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePR.R
\name{computePRcurve}
\alias{computePRcurve}
\title{Compute Precision-Recall curve}
\usage{
computePRcurve(thisModel, pres, bkg)
}
\arguments{
\item{thisModel}{maxnet object. A maxnet object representing the fitted model to be used}

\item{pres}{data.frame. Env data for pres records}

\item{bkg}{data.frame. Env data for bkg records}
}
\value{
A PR curve object produced by the function \emph{pr.curve} in the package PRROC
}
\description{
Compute a Precision-Recall (PR) curve and AUC value for a maxnet model
}
