% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyStats.R
\name{energyStats}
\alias{energyStats}
\title{Compute difference between two multivariate samples}
\usage{
energyStats(
  sample1 = NULL,
  sample2 = NULL,
  edist_method = "original",
  edist_replicates = 199
)
}
\arguments{
\item{sample1}{Numeric. Matrix or data.frame with values sampled from a multivariate distribution.}

\item{sample2}{Numeric. Matrix or data.frame with a second example of values sampled from a multivariate distribution.}

\item{edist_method}{Character. Selects the method used in a call to function \link{eqdist.etest} in the package \emph{energy}.}

\item{edist_replicates}{Numeric (integer). Number of replicates used in energy statistics calculation. Default is 199 which appears to give reliable results without significant computational cost.}
}
\value{
An object of class \emph{htest} as returned by function \link{eqdist.etest}.
}
\description{
Compute difference between two multivariate samples
}
\details{
{
This function is a wrapper around the function \emph{eqdist.etest} in the package \emph{energy} designed to facilitate its orderly and robust application to testing the difference between a base sample and a second sample (typically a sub-sample) of points in a multivariate environmental space.


}
}
\examples{
\dontrun{
m1 <- matrix(runif(100), 10, 10)
df1 <- data.frame(m1)
m2 <- matrix(runif(100), 10, 10)
df2 <- data.frame(m2)
ans1 <- energyStats(m1, m2)
print(ans1)


ans2 <- energyStats(df1, df2)
print(ans2)

m2 <- matrix(runif(120), 10, 12)
df2 <- data.frame(m2)
ans <- energyStats(df1, df2)
print(ans)
}
}
