% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrAnalysis.R
\name{envCorrAnalysis}
\alias{envCorrAnalysis}
\title{Exploratory Correlation Analysis of Environmental data}
\usage{
envCorrAnalysis(
  taxon = "",
  titleText = NULL,
  envDataPath,
  occData,
  xVar = NULL,
  yVar = NULL,
  threshold = 0.7,
  outFile = NULL,
  outPath = NULL
)
}
\arguments{
\item{taxon}{Character string.}

\item{titleText}{Character string.}

\item{envDataPath}{Character string.}

\item{occData}{Data.frame or matrix.}

\item{xVar}{Character. Name of a variable in \emph{occData} which is interpreted as the x-coordinate. If NULL (default) a search is made for nearest match to 'longitude'.}

\item{yVar}{Character. Name of a variable in \emph{occData} which is interpreted as the y-coordinate. If NULL (default) a search is made for nearest match to 'latitude'.}

\item{outFile}{Character. A non-NULL value is used as a file name to save the graphical output as a PNG file. By default, the output is plotted to the default graphics device.}

\item{outPath}{Character. Path used by ggsave in combination with outFile to save the plot.}
}
\value{
A character matrix listing the names of variables with absolute value of correlations greater than \emph{threshold} which may be candidates for removal, and the number of threshold-exceeding correlations in which a listed variable has been found.
}
\description{
Exploratory Correlation Analysis of Environmental data
}
