% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_interaction_heatmap.R
\name{interaction_heatmap}
\alias{interaction_heatmap}
\title{Plot Variable Interaction Heatmap}
\usage{
interaction_heatmap(
  feature_summary_table = "",
  num_repl = NULL,
  factor_ordering = NULL,
  min_colour = "white",
  max_colour = "blue"
)
}
\arguments{
\item{feature_summary_table}{Data.frame. Data table produced by function \link{feature_summary}}

\item{num_repl}{Integer. Number of replicate models used to prepare feature_summary_table}

\item{factor_ordering}{Character array. Variable names in an order preferred for plotting. Default of NULL uses the}

\item{min_colour}{Colour name or hexadecimal colour value. Colour for the start of the colour gradient used to shade heatmap grid cells. Default is "white"}

\item{max_colour}{Colour name or hexadecimal colour value. Colour for the upper level of the colour gradient. Default is "blue"}
}
\value{
A ggplot2 object
}
\description{
Plot Variable Interaction Heatmap
}
\details{
The data provided in the parameter feature_summary_table is the number of replicate models in which each feature was retained (ie was given a non-zero coefficient).

This does not represent feature importance. Some indirect inference about this may come from considering the importance of each of the two variables contributing to a feature as reported by \link{varImportance}.

It might be possible to compute an estimate of feature importance using the same approach implemented in \link{varImportance}.

This function is based on code found here: https://r-graph-gallery.com/79-levelplot-with-ggplot2.html
}
\examples{
\dontrun{}

}
