% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kFoldSet.R
\name{makeFolds}
\alias{makeFolds}
\title{Make k sets of indices to samples for k-fold cross-validation}
\usage{
makeFolds(numSamples, k, testSplit = 0.2)
}
\arguments{
\item{numSamples}{Integer. The number of samples to be partitioned into k folds.}

\item{k}{Integer. The number of folds.}

\item{testSplit}{Numeric. The fraction of samples to be used in making the test set}
}
\value{
A list of numeric vectors of length k, where the elements of each vector are the indices of the samples forming the TEST set for that fold.
}
\description{
The function randomly allocates sample indices to k numeric vectors. The length of each vector is determined as the nearest integer to the number of samples times the test fraction.
}
\examples{
\dontrun{ }
}
