% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectMaxnet.R
\name{projectMaxnet}
\alias{projectMaxnet}
\title{Project maxnet model}
\usage{
projectMaxnet(
  taxonName = NULL,
  maxnetModel,
  type = "exponential",
  baseOutputPath,
  fileLabel = NULL,
  makeTaxonFolder = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{taxonName}{String. Taxonomic name associated with this model; used to construct a file name for the output raster file.}

\item{maxnetModel}{String. \emph{Full} path name to the .Rd file storing a fitted maxnet model produced by the companion functions \link{fit_maxnet} and \link{fitModels}.}

\item{baseOutputPath}{String. Full path to the output folder to receive the produced raster.}

\item{fileLabel}{String. An identifying tag to be included in the output filename.}

\item{makeTaxonFolder}{Logical. Should a sub-folder on \emph{baseOutputPath} be created using \emph{taxonName}?}

\item{quiet}{Logical. Procede without emitting progress messages?}

\item{type.}{String. The type of scaling applied to predicted model values.}
}
\value{
Nothing
}
\description{
Project a maxnet model fitted using fitModel
}
\details{
{
Function \link{prepProjData} \strong{MUST} be run before calling this function to prepare global data objects needed for the projection.}
}
\examples{
\dontrun{}
}
