% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectMaxnet.R
\name{projectMaxnet}
\alias{projectMaxnet}
\title{Project maxnet model}
\usage{
projectMaxnet(
  taxonName = NULL,
  maxnetModel,
  type = "exponential",
  doClamp = FALSE,
  baseOutputPath,
  fileLabel = NULL,
  makeTaxonFolder = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{taxonName}{String. Taxonomic name associated with this model; used to construct a file name for the output raster file.}

\item{maxnetModel}{String. \emph{Full} path name to the .Rd file storing a fitted maxnet model produced by the companion functions \link{fit_maxnet} and \link{fitModels}.}

\item{type}{String. The type of scaling applied to predicted model values.}

\item{doClamp}{Logical. Should values of predictors (covariates) be clamped to those seen during model fitting? Default is FASLE, no clamping.}

\item{baseOutputPath}{String. Full path to the output folder to receive the produced raster.}

\item{fileLabel}{String. An identifying tag to be included in the output filename.}

\item{makeTaxonFolder}{Logical. Should a sub-folder on \emph{baseOutputPath} be created using \emph{taxonName}?}

\item{quiet}{Logical. Proceed without emitting progress messages?}
}
\value{
Nothing
}
\description{
Project a maxnet model fitted using fitModel
}
\details{
{
Function \link{prepProjData} \strong{MUST} be run before calling this function to prepare global data objects needed for the projection.

\strong{NOTE:} The resulting raster is written to the specified output folder with the file name composed by concatenating hte taxon name spaces repalced by underscores), "projection" and, if supplied, the character object \emph{fileTag}.

\strong{If a file with this name already exists in the output folder, it will be overwritten.}
}
}
\examples{
\dontrun{}
}
