% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDuplicates.R
\name{removeDuplicates}
\alias{removeDuplicates}
\title{Remove duplicate occurrence records}
\usage{
removeDuplicates(data = NULL, baseGrid = NULL, byGrid = FALSE, quiet = FALSE)
}
\arguments{
\item{data}{A data.frame or matrix with at least a latitude and longitude column.}

\item{baseGrid}{A RasterLayer or SpatRaster object representing the grid for enforcing one record per cell when \emph{byGrid} = TRUE.}

\item{byGrid}{Logical. If TRUE then use the rasterLayer passed in \emph{baseGrid} to select opne record per cell. If FALSE (default) then ignore anything passed in \emph{baseGrid} and remove exact duplicates based on there coordinates.}

\item{quiet}{Logical. If TRUE then emit some progress message. If FALSE (default) procede quietly.}
}
\value{
A version of \emph{data} with duplicates removed
}
\description{
Remove duplicate occurrence records by exact coordinate matching or by ensuring only one record per grid cell of a raster. Coordinates are assumed be geographic coordinates (i.e. latitude and longitude) on the WGS84 datum
}
\examples{
\dontrun{}
}
