% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsePlots.R
\name{responsePlot}
\alias{responsePlot}
\title{Response plots for maxnet-produced MaxEnt models}
\usage{
responsePlot(
  thisModel,
  variable = "all",
  importance_threshold = NULL,
  responseType = c("link", "exponential", "cloglog", "logistic"),
  plot_ylimits = c(0, 1),
  ylabel = NULL,
  filename = NULL,
  occSWD = NULL,
  bkgSWD = NULL
)
}
\arguments{
\item{thisModel}{A maxnet model object.}

\item{variable}{Character. Names of one or more variables present in the model object. The default of 'all' causes all variables to included.}

\item{importance_threshold}{Numeric. Percentage value to serve as a cut-off for variables to be included.}

\item{responseType}{Character. Response type (scaling of model output) to be used.}

\item{plot_ylimits}{Numeric vector. Limits to be used on the plot y-axis.}

\item{ylabel}{Character. Label to be used on the \emph{y}-axis replacing the default value formed from the prefix "Response" and the value of \emph{responseType}.}

\item{filename}{Character. Filename (with full path) into which graphics will be plotted.}

\item{occSWD}{Data.frame. Environmental data at occurrence locations in SWD format; used to compute variable importance.}

\item{bkgSWD}{Data.frame. Environmental data at background locations in SWD format; used to compute variable importance.}
}
\value{
A named list of ggplot graphics objects ordered in decreasing variable importance, plus the side-effect of a pdf of plots.
}
\description{
Response plots are produced for some or all variables in the supplied maxnet model object.
}
\details{
{
This function is closely based on the source code for the corresponding function in the R package \pkg{maxnet} and draws inspiration from the response plots produced for Boosted Regression Tree (BRT) models produced by the package \pkg{gbm}.

Features of this function include:
\itemize{
\item Returns a list of ggplot objects allowing individual plots to be extracted and used for other purposes
\item Following the approach in package \pkg{gbm}, incorporation of variable importance scores supplied by the companion function \link{varImportance}
\item Ability to produce a plot for \emph{all} variables in the maxnet model or a selection
}
}
}
\examples{
\dontrun{}

}
