% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responsePlots.R
\name{responsePlot}
\alias{responsePlot}
\title{Response plots for maxnet-produced MaxEnt models}
\usage{
responsePlot(
  thisModel,
  variable = "all",
  responseType = "",
  ylimits = c(0, 1),
  ylabel = NULL,
  filename = NULL,
  occSWD = NULL,
  bkgSWD = NULL
)
}
\arguments{
\item{thisModel}{A maxnet model object.}

\item{variable}{Character. Names of one or more variables present in the model object.}

\item{responseType}{Character. Response type (scaling of model output) to be used.}

\item{ylimits}{Numeric vector. Limits to be used on the plot y-axis.}

\item{ylabel}{Character. Label to be used on the y-axis replacing the default value.}

\item{filename}{Character. Filename (with full path) into which graohics will be plotted.}

\item{occSWD}{Data.frame. Environmental data at occurrence locations in SWD format; used to compute variable importance.}

\item{bkgSWD}{Data.frame. Environmental data at background locations in SWD format; used to compute variable importance.}
}
\value{
A named list of ggplot graphics objects order in decreasing varaible importance.
}
\description{
Response plots for maxnet-produced MaxEnt models
}
