% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocplot.R
\name{smoothdist}
\alias{smoothdist}
\title{Smoothed density distribution}
\usage{
smoothdist(pred, res, smoothing_df = 4)
}
\arguments{
\item{pred}{Numeric. Vector of model predictions for combined occurrence and background points.}

\item{res}{Numeric. Vector expected presence and background values (referred in the function as 'res' = 'response variable'); 0 = expected background, 1 = expected presence.}

\item{smoothing_df}{Numeric. Degrees of freedom for a natural cubic spline fitted to model predictions at combined occurrence and background locations.}
}
\value{
A data.frame with three columns:
\describe{
\item{x}{The values used to fit the smotthing function}
\item{y}{The y-values of the computed smoothing function}
\item{se}{The standard error of y-values of the computed smoothing function}
}
}
\description{
A function to generate a smoothed density function of observed probability of presence
}
\details{
{
This is an adaptation of code published by Phillips and Elith (2010. POC plots: calibrating species distribution models with presence-only data. Ecology 91:2476–2484).

It is normally called by the function \link{POCplot}.

The original script used a default smoothing degrees of freedom (\emph{smoothing_df}) of 6, but trials during development of this implementation suggest that a better default value is 4.
}
}
