% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinningReview.R
\name{thinningReview}
\alias{thinningReview}
\title{Review impact of a sequence of thinning distances}
\usage{
thinningReview(
  taxon = "",
  occData = NULL,
  envDataPath = NULL,
  outPath = "",
  threshold = 0.9,
  numReplicates = 5,
  doPlots = FALSE,
  writeResults = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{taxon}{Character. A taxon name to label output files and plots}

\item{occData}{Data frame of occurrence locations, or a character object (string) giving the path to this file. Location coordinates are expected to be in decimal degrees. An attempt is made to identify the columns storing latitude and longitude values by partila matching on "longitude" and "latitude"}

\item{envDataPath}{Characater. Full path to the raster layers of environmental data to nbe used to acssess environmental coverage. Any file format accepted by the raster package can be used, and alll files in the set must have the same gridcell size, origin and resolution}

\item{outPath}{Character. A path to a folder into which out will be written}

\item{threshold}{Numeric. A value between 0 and 1 representing the fraction of environmental space which must be covered by the thinning operation}

\item{numReplicates}{Integer. How many repeated thinning runs should be performed at each thinning distance?}

\item{doPlots}{Logical. Should plots of results for each replicate and distance combination be produced in additional to a final summary plot?}

\item{writeResults}{Logical. Should a results summary table be saved?}

\item{quiet}{Logical. Should progress messages be written to the console}
}
\value{
The largest distance in the standard sequence for which environmental coverage is >= threshold
}
\description{
The function applies a standard sequence of thinning distances (in kilometers) to occupancy records and determines the maximum distance for thinning by examining the imapct of each prospective tinnning distance on coverage of environmental conditions.
}
