% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImportance.R
\name{varImportance}
\alias{varImportance}
\title{Variable importance for a maxnet model}
\usage{
varImportance(
  theModel,
  occSWD = NULL,
  bkgSWD = NULL,
  responseType = c("link", "exponential", "cloglog", "logistic"),
  numReplicates = 5
)
}
\arguments{
\item{theModel}{Object of class maxnet}

\item{occSWD}{Data.frame. Environmental data at occurrence locations in SWD format.}

\item{bkgSWD}{Data.frame. Background environmental data in SWD format.}

\item{responseType}{Character. A MaxEnt response scale; one of "link", "exponential", "logistic" or "cloglog".}

\item{numReplicates}{Integer. Number of permutations performed to compute importance.}
}
\value{
A named vector of percent importance scores for the variables present in the maxnet model object sorted from highest to lowest.
}
\description{
Computes variable importance scores for all variables present in a maxnet model object. Scores are indicative and should be interpreted with caution as variable interaction is not considered in the computation of the scores.
}
\details{
{
The method used to compute variable importance follows that used in R packages \pkg{biomod2} and \pkg{ecospat}. First, model predictions are made for each row of the combined environmental data table formed by stacking \emph{occSWD} and \emph{bkgSWD}. This is the reference or full-model result.

For each variable in the maxnet model object, values for the variable are permuted between rows and a model prediction made for each row using the permuted or shuffled data table. The permutation is performed \emph{numReplicates} times for each variable.

At each permutation, a Pearson correlation is computed between reference predictions and the predicted values from the shuffled table. The importance score is 1 - correlation coefficient.

A vector of mean scores for each variable expressed as a percentage of the sum of all mean scores is returned.
}
}
\examples{
\dontrun{}

}
